celeste/requirements.txt
absl-py
imageio
matplotlib
moviepy
pillow
pygame==2.0.0.dev3
scipy
tensorboardX
tensorflow-gpu==1.14
-e ./third_party/atari-py
-e ./third_party/baselines
